show score for latest assessment attempts
